Adjusted the writing of the JSON file and added an error interception when copying the file and pasting it to the same location again

+++ DataProcessingService(Task1Radency)/Classes/Models/City.cs
@@ -11,7 +11,7 @@ namespace DataProcessingService_Task1Radency_.Classes.Models
     {
         public string Name { get; set; }
         public List<Service> Services { get; set; }
-        public int Total { get; set; } = 0;
+        public double Total { get; set; } = 0;
 
         public City(string name, List<Service> services)
         {
